Mark arguments as volatile and therefore requiring type guards on binary operations. This was a regression in the last release, uncovered by the PGC tests

+++ src/pyjion/ipycomp.h
@@ -428,7 +428,7 @@ public:
     virtual void emit_trace_exception() = 0;
     virtual void emit_profile_frame_entry() = 0;
     virtual void emit_profile_frame_exit() = 0;
-    virtual void emit_pgc_probe(int curByte, int stackSize) = 0;
+    virtual void emit_pgc_probe(size_t curByte, size_t stackSize) = 0;
 
     /* Compiles the generated code */
     virtual JittedCode* emit_compile() = 0;
